dashboard/requirements.txt
dash>=2.0.0
plotly>=5.0.0
pandas>=1.3.0
numpy>=1.21.0
websocket-client>=1.3.1
flask-socketio>=5.3.0
eventlet>=0.33.0  # Required for WebSocket support